requirements.txt
joblib==1.1.1
pandas==2.0.1
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.24.1
